Align message property indentation across multiple DTOs and update test cases for consistency

+++ src/database/gameplay-postgresql/animal.entity.ts
@@ -8,57 +8,57 @@ import { ProductEntity } from "./product.entity"
 @Entity("animals")
 export class AnimalEntity extends ReadableAbstractEntity {
     @Field(() => Int)
-    @Column({ name: "yield_time", type: "bigint" })
-        yieldTime: number
+    @Column({ name: "yield_time", type: "int" })
+    yieldTime: number
 
     @Field(() => Int)
-    @Column({ name: "offspring_price", type: "bigint" })
-        offspringPrice: number
+    @Column({ name: "offspring_price", type: "int" })
+    offspringPrice: number
 
     @Field(() => Boolean)
     @Column({ name: "is_nft", type: "boolean" })
-        isNFT: boolean
+    isNFT: boolean
 
     @Field(() => Int)
-    @Column({ name: "growth_time", type: "bigint" })
-        growthTime: number
+    @Column({ name: "growth_time", type: "int" })
+    growthTime: number
 
     @Field(() => Boolean)
     @Column({ name: "available_in_shop", type: "boolean" })
-        availableInShop: boolean
+    availableInShop: boolean
 
     @Field(() => Int)
-    @Column({ name: "hunger_time", type: "bigint" })
-        hungerTime: number
+    @Column({ name: "hunger_time", type: "int" })
+    hungerTime: number
 
     @Field(() => Int)
     @Column({ name: "min_harvest_quantity", type: "int" })
-        minHarvestQuantity: number
+    minHarvestQuantity: number
 
     @Field(() => Int)
     @Column({ name: "max_harvest_quantity", type: "int" })
-        maxHarvestQuantity: number
+    maxHarvestQuantity: number
 
     @Field(() => Int)
-    @Column({ name: "basic_harvest_experiences", type: "bigint" })
-        basicHarvestExperiences: number
+    @Column({ name: "basic_harvest_experiences", type: "int" })
+    basicHarvestExperiences: number
 
     @Field(() => Int)
-    @Column({ name: "premium_harvest_experiences", type: "bigint" })
-        premiumHarvestExperiences: number
+    @Column({ name: "premium_harvest_experiences", type: "int" })
+    premiumHarvestExperiences: number
 
     @Field(() => AnimalType)
     @Column({ name: "type", type: "enum", enum: AnimalType })
-        type: AnimalType
+    type: AnimalType
 
     @Field(() => Float)
     @Column({ name: "sick_chance", type: "float" })
-        sickChance: number
+    sickChance: number
 
     @OneToOne(() => ProductEntity, { cascade: true, onDelete: "CASCADE" })
     @JoinColumn({
         name: "product_id",
         referencedColumnName: "id"
     })
-        product: ProductEntity
+    product: ProductEntity
 }
